+++ src/main/java/org/sophy/sophy/domain/other/ScheduledBooktalk.java
@@ -1,4 +1,4 @@
-package org.sophy.sophy.domain;
+package org.sophy.sophy.domain.other;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
+import org.sophy.sophy.domain.Booktalk;
 
 @Entity
 @Getter
