+++ src/main/java/org/sophy/sophy/service/SchedulerService.java
@@ -5,7 +5,7 @@ import java.util.Arrays;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.sophy.sophy.domain.Booktalk;
-import org.sophy.sophy.domain.ScheduledBooktalkConverter;
+import org.sophy.sophy.domain.other.ScheduledBooktalkConverter;
 import org.sophy.sophy.domain.enumerate.BooktalkStatus;
 import org.sophy.sophy.infrastructure.query.BooktalkQueryRepository;
 import org.sophy.sophy.service.api.BooktalkService;
